+++ examples/tracing/outputs.tf
@@ -1,0 +1,4 @@
+output "lambda_arn" {
+  value       = module.lambda.arn
+  description = "ARN of the given lambda."
+}